% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champ.normm.R
\name{champ.normm}
\alias{champ.normm}
\title{champ.normm}
\usage{
champ.normm(
  beta = myLoad$beta,
  rgSet = myLoad$rgSet,
  mset = myLoad$mset,
  resultsDir = "./CHAMP_Normalization/",
  method = "BMIQ",
  plotBMIQ = FALSE,
  arraytype = "mouse",
  cores = 4
)
}
\arguments{
\item{beta}{Beta Matrix}

\item{rgSet}{RGset (optional)}

\item{mset}{Mset (optional)}

\item{cores}{used cores, default = 4}
}
\value{
Normalised beta matrix
}
\description{
ChAMP normalisation for the Illumina Mouse Methylation array. Like the ChAMP champ.norm package, it can use SWAN, funNorm, PBC, or BMIQ normalisation. Depending on normalisation algorithm, different inputs may be needed (beta matrix, rgSet, mset). Please refer to the ChAMP package manual.
}
